refactor: split main component into smaller files
