+++ src/components/TetrisButton/index.ts
@@ -1,1 +1,2 @@
 export { default as TetrisButton } from './TetrisButton';
+export type { TetrisButtonProps } from './interface';
